+++ src/components/MusicInfo/Vinyl/Album.tsx
