+++ src/components/MusicInfo/Vinyl/Cover.tsx
