+++ src/components/MusicInfo/Vinyl/Print.tsx
