+++ src/components/MusicInfo/Vinyl/VinylBody.tsx
