+++ src/components/MusicInfo/_shared/FixedWrap.tsx
